fix bf16

+++ src/deepforest/utilities.py
@@ -44,7 +44,6 @@ def load_config(
     Returns:
         config (DictConfig): composed configuration
     """
-
     if not config_name.endswith(".yaml"):
         config_name += ".yaml"
 
@@ -375,7 +374,6 @@ def format_geometry(predictions, scores=True, geom_type=None):
         df: a pandas dataframe
         None if the dataframe is empty
     """
-
     # Detect geometry type
     if geom_type is None:
         geom_type = determine_geometry_type(predictions)
@@ -409,13 +407,13 @@ def format_keypoints(prediction, scores=True):
         return None
 
     df = pd.DataFrame(
-        prediction["points"].cpu().detach().numpy(),
+        prediction["points"].cpu().detach().float().numpy(),
         columns=["x", "y"],
     )
     df["label"] = prediction["labels"].cpu().detach().numpy()
 
     if scores:
-        df["score"] = prediction["scores"].cpu().detach().numpy()
+        df["score"] = prediction["scores"].cpu().detach().float().numpy()
 
     df["geometry"] = df.apply(lambda x: shapely.geometry.Point(x.x, x.y), axis=1)
     return df
@@ -792,7 +790,6 @@ def geo_to_image_coordinates(gdf, image_bounds, image_resolution):
     Returns:
         gdf: a geopandas dataframe with the transformed to image origin. CRS is removed
     """
-
     if len(image_bounds) != 4:
         raise ValueError("image_bounds must be a tuple of (left, bottom, right, top)")
 
@@ -811,7 +808,6 @@ def geo_to_image_coordinates(gdf, image_bounds, image_resolution):
 
 def round_with_floats(x):
     """Check if string x is float or int, return int, rounded if needed."""
-
     try:
         result = int(x)
     except BaseException:
